chore: add /lib in client for contexts & types
- fixed vite fast refresh (no mixed exports)
